Fix usage of frailty terms in survreg() dummy examples
survival 3.2-9 now errors when this is used incorrectly. Their NEWS bullet says:

"Add an explicit error message for survreg() with a frailty term. This usage was never supported – it is integrating out the appropriate density for a coxph model, which is the wrong one for survreg – but because the code would run we've found that people assumed it must be right, despite a statement to the contrary in the survival vignette. Long overdue fix."